perf: fix elevation of privilege vulnerability
An elevation of privilege vulnerability in the Qualcomm performance component
could enable a local malicious application to execute arbitrary code within the
context of the kernel. This issue is rated as Critical due to the possibility
of a local permanent device compromise, which may require reflashing the
operating system to repair the device.

Change-Id: I3c0064f9a832d972b3e09bdab95cbdce3fdeaa58
Android: A_28086229
Issue: FEIJAO-1614